+++ Core/HLE/__sceAudio.cpp
@@ -355,7 +355,7 @@ int __AudioMix(short *outstereo, int numFrames)
 	if (remains > 0)
 		memset(outstereo + numFrames * 2 - remains, 0, remains);
 
-	if (sz1 + sz2 < numFrames) {
+	if (sz1 + sz2 < (size_t)numFrames) {
 		underrun = (int)(sz1 + sz2) / 2;
 		DEBUG_LOG(HLE, "Audio out buffer UNDERRUN at %i of %i", underrun, numFrames);
 	}
